requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djoser==2.2.2
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1